update track page animation
