+++ app/routes/_app.track.add/_index.tsx
@@ -1,6 +1,7 @@
 import * as HeadlessUI from '@headlessui/react'
 import * as RemixNode from '@remix-run/node'
 import * as RemixReact from '@remix-run/react'
+import { AnimatePresence, motion } from 'framer-motion'
 import * as React from 'react'
 import * as RemixParamsHelper from 'remix-params-helper'
 import * as Zod from 'zod'
@@ -27,7 +28,7 @@ export async function action({ request }: RemixNode.ActionArgs) {
 
   const result = await RemixParamsHelper.getFormData(request, FormDataSchema)
   if (!result.success) {
-    return RemixNode.json({ errors: result.errors }, { status: 400 })
+    return RemixNode.json({ ok: false, errors: result.errors }, { status: 400 })
   }
 
   const { name, ...progression } = result.data
@@ -41,7 +42,7 @@ export async function action({ request }: RemixNode.ActionArgs) {
 
   const errors = UtilsServer.Character.validateAscensionRequirement(progression)
   if (errors) {
-    return RemixNode.json({ errors }, { status: 400 })
+    return RemixNode.json({ ok: false, errors }, { status: 400 })
   }
 
   await CharacterModel.upsertCharacterTrack({
@@ -50,7 +51,7 @@ export async function action({ request }: RemixNode.ActionArgs) {
     accountId,
   })
 
-  return RemixNode.redirect('/track')
+  return RemixNode.json({ ok: true, errors: {} })
 }
 
 export async function loader({ request }: RemixNode.LoaderArgs) {
@@ -67,43 +68,41 @@ export async function loader({ request }: RemixNode.LoaderArgs) {
 
 export default function AddTrackPage() {
   const { nonTrackCharacterNames } = RemixReact.useLoaderData<typeof loader>()
-  const actionData = RemixReact.useActionData<typeof action>()
+
+  const submitFetcher = RemixReact.useFetcher<{
+    ok: boolean
+    errors: {
+      [key: string]: string
+    }
+  }>()
 
   const navigate = RemixReact.useNavigate()
-  const transition = RemixReact.useTransition()
-  const busy = transition.state === 'submitting'
+  const busy = submitFetcher.state === 'submitting'
 
-  const [open, setOpen] = React.useState(false)
+  const [open, setOpen] = React.useState(true)
+
   React.useEffect(() => {
-    setOpen(true)
-  }, [])
+    if (submitFetcher.data?.ok) {
+      setOpen(false)
+    }
+  }, [submitFetcher.data])
 
-  const timerRef = React.useRef<NodeJS.Timeout>()
-  function handleClose() {
-    if (timerRef.current) clearTimeout(timerRef.current)
-    setOpen(false)
-
-    const timer = setTimeout(() => {
-      navigate('/track')
-    }, 500)
-    timerRef.current = timer
-  }
-
-  const fetcher = RemixReact.useFetcher<CharacterTypes.CharacterProgression>()
+  const progressionfetcher =
+    RemixReact.useFetcher<CharacterTypes.CharacterProgression>()
 
   function handleFetchProgression(name: string) {
     if (name === '') return
-    fetcher.load('/api/get-character-progression?name=' + name)
+    progressionfetcher.load('/api/get-character-progression?name=' + name)
   }
 
-  const character = fetcher.data
+  const character = progressionfetcher.data
 
   return (
     <HeadlessUI.Transition.Root show={open} as={React.Fragment}>
       <HeadlessUI.Dialog
         as="div"
         className="relative z-10"
-        onClose={handleClose}
+        onClose={() => setOpen(false)}
       >
         <HeadlessUI.Transition.Child
           as={React.Fragment}
@@ -120,148 +119,168 @@ export default function AddTrackPage() {
         <div className="fixed inset-0 overflow-hidden">
           <div className="absolute inset-0 overflow-hidden">
             <div className="pointer-events-none fixed inset-y-0 right-0 flex max-w-full pl-10 sm:pl-16">
-              <HeadlessUI.Transition.Child
-                as={React.Fragment}
-                enter="transform transition ease-in-out duration-500 sm:duration-700"
-                enterFrom="translate-x-full"
-                enterTo="translate-x-0"
-                leave="transform transition ease-in-out duration-500 sm:duration-700"
-                leaveFrom="translate-x-0"
-                leaveTo="translate-x-full"
+              <AnimatePresence
+                exitBeforeEnter
+                onExitComplete={() => navigate('/track')}
               >
-                <HeadlessUI.Dialog.Panel className="pointer-events-auto w-screen max-w-2xl">
-                  <RemixReact.Form
-                    method="post"
-                    className="flex h-full flex-col overflow-y-scroll bg-gray-2 shadow-xl"
+                {open && (
+                  <motion.div
+                    className="flex max-w-full"
+                    transition={{
+                      x: { duration: 0.5 },
+                      default: { ease: 'easeInOut' },
+                    }}
+                    initial={{ x: '100%' }}
+                    animate={{ x: 0 }}
+                    exit={{ x: '100%' }}
                   >
-                    <div className="flex-1">
-                      <div className="bg-gray-3 px-4 py-6 sm:px-6">
-                        <div className="flex items-start justify-between space-x-3">
-                          <HeadlessUI.Dialog.Title className="text-lg font-medium text-gray-12">
-                            Track a character
-                          </HeadlessUI.Dialog.Title>
+                    <HeadlessUI.Dialog.Panel className="pointer-events-auto w-screen max-w-2xl">
+                      <submitFetcher.Form
+                        method="post"
+                        className="flex h-full flex-col overflow-y-scroll bg-gray-2 shadow-xl"
+                      >
+                        <div className="flex-1">
+                          <div className="bg-gray-3 px-4 py-6 sm:px-6">
+                            <div className="flex items-start justify-between space-x-3">
+                              <HeadlessUI.Dialog.Title className="text-lg font-medium text-gray-12">
+                                Track a character
+                              </HeadlessUI.Dialog.Title>
 
-                          <div className="flex h-7 items-center">
-                            <Button.Icon onClick={handleClose}>
-                              <span className="sr-only">Close panel</span>
-                              <Icon.Outline
-                                name="x"
-                                className="h-6 w-6"
-                                aria-hidden="true"
+                              <div className="flex h-7 items-center">
+                                <Button.Icon onClick={() => setOpen(false)}>
+                                  <span className="sr-only">Close panel</span>
+                                  <Icon.Outline
+                                    name="x"
+                                    className="h-6 w-6"
+                                    aria-hidden="true"
+                                  />
+                                </Button.Icon>
+                              </div>
+                            </div>
+                          </div>
+
+                          <div className="space-y-6 py-6 sm:space-y-0 sm:divide-y sm:divide-gray-6 sm:py-0">
+                            {nonTrackCharacterNames.length > 0 ? (
+                              <Combobox
+                                options={nonTrackCharacterNames}
+                                fetchProgressionHandler={handleFetchProgression}
                               />
-                            </Button.Icon>
+                            ) : (
+                              <div className="mt-12 px-4 text-center sm:px-6">
+                                <h3 className="text-lg font-medium leading-6 text-gray-12">
+                                  No trackable character
+                                </h3>
+                                <p className="text-gray-11">
+                                  You don't have any characters that can be
+                                  tracked. This means that you already have all
+                                  characters maxed out and/or tracked.
+                                </p>
+                              </div>
+                            )}
+
+                            {character && (
+                              <>
+                                <div className="px-4 sm:px-6 sm:py-5">
+                                  <p className="text-gray-11">
+                                    Last updated progression for{' '}
+                                    <span className="text-gray-12">
+                                      {character.name}
+                                    </span>
+                                  </p>
+                                </div>
+
+                                <ProgressionField
+                                  label="Level"
+                                  name="level"
+                                  currentValue={character.progression.level}
+                                  min={character.progression.level}
+                                  max={90}
+                                  error={submitFetcher.data?.errors?.level}
+                                />
+
+                                <ProgressionField
+                                  label="Ascension"
+                                  name="ascension"
+                                  currentValue={character.progression.ascension}
+                                  min={character.progression.ascension}
+                                  max={6}
+                                  error={submitFetcher.data?.errors?.ascension}
+                                />
+
+                                <ProgressionField
+                                  id="normal-attack"
+                                  label="Normal Attack"
+                                  name="normalAttack"
+                                  currentValue={
+                                    character.progression.normalAttack
+                                  }
+                                  min={character.progression.normalAttack}
+                                  max={10}
+                                  error={
+                                    submitFetcher.data?.errors?.normalAttack
+                                  }
+                                />
+
+                                <ProgressionField
+                                  id="elemental-skill"
+                                  label="Elemental Skill"
+                                  name="elementalSkill"
+                                  currentValue={
+                                    character.progression.elementalSkill
+                                  }
+                                  min={character.progression.elementalSkill}
+                                  max={10}
+                                  error={
+                                    submitFetcher.data?.errors?.elementalSkill
+                                  }
+                                />
+
+                                <ProgressionField
+                                  id="elemental-burst"
+                                  label="Elemental Burst"
+                                  name="elementalBurst"
+                                  currentValue={
+                                    character.progression.elementalBurst
+                                  }
+                                  min={character.progression.elementalBurst}
+                                  max={10}
+                                  error={
+                                    submitFetcher.data?.errors?.elementalBurst
+                                  }
+                                />
+                              </>
+                            )}
                           </div>
                         </div>
-                      </div>
 
-                      <div className="space-y-6 py-6 sm:space-y-0 sm:divide-y sm:divide-gray-6 sm:py-0">
-                        {nonTrackCharacterNames.length > 0 ? (
-                          <Combobox
-                            options={nonTrackCharacterNames}
-                            fetchProgressionHandler={handleFetchProgression}
-                          />
-                        ) : (
-                          <div className="mt-12 px-4 text-center sm:px-6">
-                            <h3 className="text-lg font-medium leading-6 text-gray-12">
-                              No trackable character
-                            </h3>
-                            <p className="text-gray-11">
-                              You don't have any characters that can be tracked.
-                              This means that you already have all characters
-                              maxed out and/or tracked.
-                            </p>
+                        <div className="flex-shrink-0 border-t border-gray-6 px-4 py-5 sm:px-6">
+                          <div className="flex justify-end space-x-3">
+                            <Button.Base
+                              type="button"
+                              variant="basic"
+                              disabled={busy}
+                              onClick={() => setOpen(false)}
+                            >
+                              {nonTrackCharacterNames.length > 0
+                                ? 'Cancel'
+                                : 'Close'}
+                            </Button.Base>
+                            {nonTrackCharacterNames.length > 0 && (
+                              <Button.Base
+                                id="track"
+                                type="submit"
+                                disabled={busy}
+                              >
+                                {busy ? 'Tracking...' : 'Track'}
+                              </Button.Base>
+                            )}
                           </div>
-                        )}
-
-                        {character && (
-                          <>
-                            <div className="px-4 sm:px-6 sm:py-5">
-                              <p className="text-gray-11">
-                                Last updated progression for{' '}
-                                <span className="text-gray-12">
-                                  {character.name}
-                                </span>
-                              </p>
-                            </div>
-
-                            <ProgressionField
-                              label="Level"
-                              name="level"
-                              currentValue={character.progression.level}
-                              min={character.progression.level}
-                              max={90}
-                              error={actionData?.errors?.level}
-                            />
-
-                            <ProgressionField
-                              label="Ascension"
-                              name="ascension"
-                              currentValue={character.progression.ascension}
-                              min={character.progression.ascension}
-                              max={6}
-                              error={actionData?.errors?.ascension}
-                            />
-
-                            <ProgressionField
-                              id="normal-attack"
-                              label="Normal Attack"
-                              name="normalAttack"
-                              currentValue={character.progression.normalAttack}
-                              min={character.progression.normalAttack}
-                              max={10}
-                              error={actionData?.errors?.normalAttack}
-                            />
-
-                            <ProgressionField
-                              id="elemental-skill"
-                              label="Elemental Skill"
-                              name="elementalSkill"
-                              currentValue={
-                                character.progression.elementalSkill
-                              }
-                              min={character.progression.elementalSkill}
-                              max={10}
-                              error={actionData?.errors?.elementalSkill}
-                            />
-
-                            <ProgressionField
-                              id="elemental-burst"
-                              label="Elemental Burst"
-                              name="elementalBurst"
-                              currentValue={
-                                character.progression.elementalBurst
-                              }
-                              min={character.progression.elementalBurst}
-                              max={10}
-                              error={actionData?.errors?.elementalBurst}
-                            />
-                          </>
-                        )}
-                      </div>
-                    </div>
-
-                    <div className="flex-shrink-0 border-t border-gray-6 px-4 py-5 sm:px-6">
-                      <div className="flex justify-end space-x-3">
-                        <Button.Base
-                          type="button"
-                          variant="basic"
-                          disabled={busy}
-                          onClick={handleClose}
-                        >
-                          {nonTrackCharacterNames.length > 0
-                            ? 'Cancel'
-                            : 'Close'}
-                        </Button.Base>
-                        {nonTrackCharacterNames.length > 0 && (
-                          <Button.Base id="track" type="submit" disabled={busy}>
-                            {busy ? 'Tracking...' : 'Track'}
-                          </Button.Base>
-                        )}
-                      </div>
-                    </div>
-                  </RemixReact.Form>
-                </HeadlessUI.Dialog.Panel>
-              </HeadlessUI.Transition.Child>
+                        </div>
+                      </submitFetcher.Form>
+                    </HeadlessUI.Dialog.Panel>
+                  </motion.div>
+                )}
+              </AnimatePresence>
             </div>
           </div>
         </div>
